requirements.txt
eze-cli
pytest==6.2.2
pytest-asyncio==0.15.0
pytest-cov==2.11.1
pytest-mock==3.5.1
pytest-snapshot==0.5.0